// src/types/blockchain-address.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * An address on a blockchain
 */
export interface BlockchainAddress {
  address: string;
  type: "eoa" | "safe" | "creator" | "factory" | "contract";
  [k: string]: unknown;
}